src/sample-mcp-server/pyproject.toml
[project]
name = "sample-mcp-server"
version = "0.1.0"
description = "MCP server exposing Setlist.fm API as tools."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "fastmcp>=2.9.2",
    "fastapi==0.115.12",
    "fastapi-mcp",
    "azure-monitor-opentelemetry",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-httpx",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-starlette"
]